requirements_pythonanywhere.txt
# PythonAnywhere için gerekli paketler
# Bu dosya PythonAnywhere'de çalıştırılacak

# Flask ve temel paketler
Flask==2.3.3
python-dotenv==1.0.0
Werkzeug==2.3.7

# HTTP istekleri için
requests==2.31.0
urllib3==2.0.7

# Twitter API için
tweepy==4.14.0

# Web scraping için
beautifulsoup4==4.12.2
lxml==4.9.3

# Email gönderimi için
secure-smtplib==0.1.5

# JSON işlemleri için (Python'da built-in, ama güvenlik için)
jsonschema==4.19.2

# Hash ve güvenlik için (Python'da built-in)
cryptography==41.0.7

# Dosya işlemleri için
Pillow==10.0.1

# Logging için (Python'da built-in, ama gelişmiş özellikler için)
colorlog==6.7.0

# Tarih/saat işlemleri için (Python'da built-in, ama gelişmiş özellikler için)
python-dateutil==2.8.2

# URL işlemleri için (Python'da built-in, ama güvenlik için)
validators==0.22.0

# Markdown işlemleri için
markdown==3.5.1

# CSV işlemleri için (Python'da built-in, ama gelişmiş özellikler için)
pandas==2.1.1

# Matematik işlemleri için
numpy==1.24.3

# Web framework ek paketleri
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1

# Güvenlik için
Flask-Talisman==1.1.0
Flask-SeaSurf==1.1.1

# Cache için
Flask-Caching==2.1.0

# Session yönetimi için
Flask-Session==0.5.0

# API rate limiting için
Flask-Limiter==3.5.0

# CORS için
Flask-CORS==4.0.0

# Debug toolbar için (sadece development)
Flask-DebugToolbar==0.13.1

# Database için (opsiyonel)
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5

# Background tasks için
celery==5.3.4

# Monitoring için
sentry-sdk[flask]==1.38.0

# Performance için
gunicorn==21.2.0
gevent==23.9.1

# Development tools (sadece development ortamında)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0 